+++ src/test/kotlin/com/dotori/v2/domain/self_study/controller/ApplySelfStudyControllerTest.kt
@@ -1,0 +1,51 @@
+package com.dotori.v2.domain.self_study.controller
+
+import com.dotori.v2.domain.self_study.presentation.councillor.CouncillorSelfStudyController
+import com.dotori.v2.domain.self_study.presentation.developer.DeveloperSelfStudyController
+import com.dotori.v2.domain.self_study.presentation.member.MemberSelfStudyController
+import com.dotori.v2.domain.self_study.service.ApplySelfStudyService
+import com.dotori.v2.domain.self_study.service.GetSelfStudyInfoService
+import io.kotest.core.spec.style.BehaviorSpec
+import io.kotest.matchers.shouldBe
+import io.mockk.every
+import io.mockk.mockk
+import io.mockk.verify
+import org.springframework.http.HttpStatus
+
+class ApplySelfStudyControllerTest : BehaviorSpec({
+    val service = mockk<ApplySelfStudyService>()
+    val getSelfStudyInfoService = mockk<GetSelfStudyInfoService>()
+    val councillorSelfStudyController = CouncillorSelfStudyController(service, getSelfStudyInfoService)
+    val developerSelfStudyController = DeveloperSelfStudyController(service, getSelfStudyInfoService)
+    val memberSelfStudyController = MemberSelfStudyController(service, getSelfStudyInfoService)
+    every { service.execute() } returns Unit
+    given("요청이 들어오면") {
+        `when`("councillorController is received") {
+            val response = councillorSelfStudyController.applySelfStudy()
+            then("서비스가 한번은 실행되어야 함") {
+                verify(exactly = 1) { service.execute() }
+            }
+            then("response status should be ok") {
+                response.statusCode shouldBe HttpStatus.OK
+            }
+        }
+        `when`("developerController is received") {
+            val response = developerSelfStudyController.applySelfStudy()
+            then("서비스가 한번은 실행되어야 함") {
+                verify(exactly = 2) { service.execute() }
+            }
+            then("response status should be ok") {
+                response.statusCode shouldBe HttpStatus.OK
+            }
+        }
+        `when`("memberController is received") {
+            val response = memberSelfStudyController.applySelfStudy()
+            then("서비스가 한번은 실행되어야 함") {
+                verify(exactly = 3) { service.execute() }
+            }
+            then("response status should be ok") {
+                response.statusCode shouldBe HttpStatus.OK
+            }
+        }
+    }
+})
